refactor: update CartContext imports to ShoppingCartContext across components
